allow keeping prunable data longer

+++ src/java/nxt/db/PrunableDbTable.java
@@ -1,5 +1,6 @@
 package nxt.db;
 
+import nxt.Constants;
 import nxt.Nxt;
 import nxt.util.Logger;
 
@@ -20,17 +21,18 @@ public abstract class PrunableDbTable<T> extends PersistentDbTable<T> {
     @Override
     public void trim(int height) {
         super.trim(height);
-        try (Connection con = db.getConnection();
-             PreparedStatement pstmt = con.prepareStatement("DELETE FROM " + table + " WHERE expiration < ?")) {
-            pstmt.setInt(1, Nxt.getEpochTime());
-            int deleted = pstmt.executeUpdate();
-            if (deleted > 0) {
-                Logger.logDebugMessage("Deleted " + deleted + " expired prunable data from " + table);
+        if (Constants.ENABLE_PRUNING) {
+            try (Connection con = db.getConnection();
+                 PreparedStatement pstmt = con.prepareStatement("DELETE FROM " + table + " WHERE expiration < ?")) {
+                pstmt.setInt(1, Nxt.getEpochTime());
+                int deleted = pstmt.executeUpdate();
+                if (deleted > 0) {
+                    Logger.logDebugMessage("Deleted " + deleted + " expired prunable data from " + table);
+                }
+            } catch (SQLException e) {
+                throw new RuntimeException(e.toString(), e);
             }
-        } catch (SQLException e) {
-            throw new RuntimeException(e.toString(), e);
         }
-
     }
 
 }
